Prepare paid App Store release

+++ Sources/DPFActivityAttributes.swift
@@ -6,8 +6,10 @@ import Foundation
 /// the extension and also surfaces the compact presentation in CarPlay.
 struct DPFActivityAttributes: ActivityAttributes {
     struct ContentState: Codable, Hashable {
-        var loadPercent: Int?
-        var regenProgressPercent: Int?
+        /// Kept as a decimal because the proprietary FCA load index can
+        /// legitimately exceed 100 and the phone dashboard shows one decimal.
+        var loadPercent: Double?
+        var regenProgressPercent: Double?
         var isRegenerating: Bool
         var exhaustTemperatureC: Int?
         var updatedAt: Date
